move to enum package
