--- src/main/utils/CurrentScene.java
@@ -1,7 +1,0 @@
-package main.utils;
-
-public enum CurrentScene {
-    START,
-    SETTINGS,
-    RANKING;
-}
