+++ src/main/utils/enums/CurrentScene.java
@@ -1,0 +1,7 @@
+package main.utils.enums;
+
+public enum CurrentScene {
+    START,
+    SETTINGS,
+    RANKING;
+}
