+++ src/main/utils/enums/EntityType.java
@@ -1,4 +1,4 @@
-package main.utils;
+package main.utils.enums;
 
 public enum EntityType {
    PLAYER,
