+++ src/main/utils/enums/Fxml.java
@@ -1,4 +1,4 @@
-package main.utils;
+package main.utils.enums;
 
 public enum Fxml {
     GAME("game"),
