+++ src/main/utils/enums/Orientations.java
@@ -1,4 +1,4 @@
-package main.utils;
+package main.utils.enums;
 
 public enum Orientations {
     HORIZONTAL,
