+++ src/main/utils/enums/SpriteType.java
@@ -1,4 +1,4 @@
-package main.utils;
+package main.utils.enums;
 
 public enum SpriteType {
     SINGLE,
