+++ src/main/utils/enums/WeaponType.java
@@ -1,4 +1,4 @@
-package main.utils;
+package main.utils.enums;
 
 public enum WeaponType {
     NPC,
